Update OLuLu_011b.py
Because of new current time input requirement, the delay added for preventing entering main() at the minute of 29 and 59 was moved to the main() loop.  The getweight() function was re-written, in accordance of the code of the Arduino side, to make the data transferred from the Arduino side to the board starting with an 'A' and ending with '\n'. Therefore, now the getweight() should be able to make sure if the entire information sending from the Arduino is received.